+++ src/__tests__/integration/control-plane/control-plane-heartbeat-mutations.test.ts
@@ -6,11 +6,7 @@ import { describe, expect, it } from 'vitest';
 import { createFileHeartbeatTaskStore, type HeartbeatTask } from '../../../index.js';
 import { ensureWorkspaceCatalog } from '../../../core/runtime/workspaces.js';
 import { controlPlaneRouter } from '../../../server/features/control-plane/router.js';
-import {
-  listControlPlaneHeartbeatTasks,
-  setControlPlaneHeartbeatTaskEnabled,
-  triggerControlPlaneHeartbeatTaskRun,
-} from '../../../server/features/control-plane/services/heartbeat.js';
+import { ControlPlaneHeartbeatController } from '../../../server/features/control-plane/controllers/heartbeat.js';
 
 function createTask(partial: Partial<HeartbeatTask> = {}): HeartbeatTask {
   return {
@@ -30,15 +26,15 @@ describe('control-plane heartbeat mutations', () => {
     const store = createFileHeartbeatTaskStore({ dir: join(stateRoot, 'heartbeat') });
     await store.saveTask(createTask());
 
-    const disabled = await setControlPlaneHeartbeatTaskEnabled(stateRoot, 'repo-check', false);
+    const disabled = await ControlPlaneHeartbeatController.setTaskEnabled(stateRoot, 'repo-check', false);
     expect(disabled).toMatchObject({ taskId: 'repo-check', enabled: false, status: 'idle' });
     expect(disabled.nextRunAt).toBeUndefined();
 
-    const enabled = await setControlPlaneHeartbeatTaskEnabled(stateRoot, 'repo-check', true);
+    const enabled = await ControlPlaneHeartbeatController.setTaskEnabled(stateRoot, 'repo-check', true);
     expect(enabled).toMatchObject({ taskId: 'repo-check', enabled: true });
     expect(enabled.nextRunAt).toBeTruthy();
 
-    const views = await listControlPlaneHeartbeatTasks(stateRoot);
+    const views = await ControlPlaneHeartbeatController.listTasks(stateRoot);
     expect(views[0]).toMatchObject({ taskId: 'repo-check', enabled: true });
   });
 
@@ -47,10 +43,10 @@ describe('control-plane heartbeat mutations', () => {
     const store = createFileHeartbeatTaskStore({ dir: join(stateRoot, 'heartbeat') });
     await store.saveTask(createTask({ enabled: false, status: 'idle', nextRunAt: undefined }));
 
-    await expect(triggerControlPlaneHeartbeatTaskRun(stateRoot, 'repo-check')).rejects.toThrow(/disabled/i);
+    await expect(ControlPlaneHeartbeatController.triggerTaskRun(stateRoot, 'repo-check')).rejects.toThrow(/disabled/i);
 
-    await setControlPlaneHeartbeatTaskEnabled(stateRoot, 'repo-check', true);
-    const triggered = await triggerControlPlaneHeartbeatTaskRun(stateRoot, 'repo-check');
+    await ControlPlaneHeartbeatController.setTaskEnabled(stateRoot, 'repo-check', true);
+    const triggered = await ControlPlaneHeartbeatController.triggerTaskRun(stateRoot, 'repo-check');
     expect(triggered).toMatchObject({ taskId: 'repo-check', enabled: true, status: 'waiting' });
     expect(triggered.nextRunAt).toBeTruthy();
