+++ src/core/chat/engine/conversation-engine.ts
@@ -1,17 +1,13 @@
 import { normalizeConversationEngineConfig } from './config.js';
-import { createConversationSessionService } from './sessions/service.js';
-import { createConversationTurnService } from './turns/service.js';
+import { FileConversationSessionService } from './sessions/service.js';
+import { EngineConversationTurnService } from './turns/service.js';
 import type { ConversationEngine, ConversationEngineConfig } from './types.js';
 
 export function createConversationEngine(config: ConversationEngineConfig): ConversationEngine {
   const normalizedConfig = normalizeConversationEngineConfig(config);
 
   return {
-    sessions: createConversationSessionService({
-      config: normalizedConfig,
-    }),
-    turns: createConversationTurnService({
-      config: normalizedConfig,
-    }),
+    sessions: new FileConversationSessionService(normalizedConfig),
+    turns: new EngineConversationTurnService(normalizedConfig),
   };
 }
