+++ src/core/chat/engine/turns/service.ts
@@ -1,63 +1,61 @@
 import { runConversationTurn, clearConversationTurnLease } from './run-conversation-turn.js';
-import { readChatSession } from '../sessions/repository/file-chat-session-repository.js';
+import { FileConversationSessionService } from '../sessions/service.js';
 import type { NormalizedConversationEngineConfig } from '../config.js';
 import type {
+  ClearConversationTurnLeaseInput,
   ContinueConversationTurnInput,
+  ConversationSessionService,
   ConversationTurnService,
   SubmitConversationTurnInput,
+  SubmitConversationTurnResult,
 } from '../types.js';
 import { normalizeConversationEngineHost } from './host.js';
 
-export function createConversationTurnService(args: {
-  config: NormalizedConversationEngineConfig;
-}): ConversationTurnService {
-  const { config } = args;
+export class EngineConversationTurnService implements ConversationTurnService {
+  private readonly sessions: ConversationSessionService;
 
-  async function submit(input: SubmitConversationTurnInput) {
+  constructor(private readonly config: NormalizedConversationEngineConfig) {
+    this.sessions = new FileConversationSessionService(config);
+  }
+
+  async submit(input: SubmitConversationTurnInput): Promise<SubmitConversationTurnResult> {
     const normalizedHost = normalizeConversationEngineHost(input.host);
     return await runConversationTurn({
-      workspaceRoot: config.workspaceRoot,
-      stateRoot: config.stateRoot,
-      sessionStoragePath: config.sessionStoragePath,
+      workspaceRoot: this.config.workspaceRoot,
+      stateRoot: this.config.stateRoot,
+      sessionStoragePath: this.config.sessionStoragePath,
       sessionId: input.sessionId,
       prompt: input.prompt,
-      apiKey: config.apiKey,
-      preferApiKey: config.preferApiKey,
-      credentialStorePath: config.credentialStorePath,
-      systemContext: config.systemContext,
-      memoryMaintenanceMode: input.memoryMaintenanceMode ?? config.memoryMaintenanceMode,
+      apiKey: this.config.apiKey,
+      preferApiKey: this.config.preferApiKey,
+      credentialStorePath: this.config.credentialStorePath,
+      systemContext: this.config.systemContext,
+      memoryMaintenanceMode: input.memoryMaintenanceMode ?? this.config.memoryMaintenanceMode,
       host: normalizedHost.turnHost,
-      approvalPolicies: input.approvalPolicies ?? config.approvalPolicies,
-      traceSummarizerRegistry: input.traceSummarizerRegistry ?? config.traceSummarizerRegistry,
+      approvalPolicies: input.approvalPolicies ?? this.config.approvalPolicies,
+      traceSummarizerRegistry: input.traceSummarizerRegistry ?? this.config.traceSummarizerRegistry,
       onCompactionStatus: normalizedHost.onCompactionStatus,
       onAssistantStream: normalizedHost.onAssistantStream,
       onTraceEvent: normalizedHost.onTraceEvent,
       abortSignal: input.abortSignal,
       leaseOwner: input.leaseOwner,
-      traceDir: config.traceDir,
+      traceDir: this.config.traceDir,
     });
   }
 
-  async function continueTurn(input: ContinueConversationTurnInput) {
-    const session = readChatSession(config.sessionStoragePath, input.sessionId, config.apiKeyPresent);
-    if (!session) {
-      throw new Error(`Chat session not found: ${input.sessionId}`);
-    }
+  async continue(input: ContinueConversationTurnInput): Promise<SubmitConversationTurnResult> {
+    const session = this.sessions.require(input.sessionId);
     if (!input.prompt && (!session.history.length || !session.lastContinuePrompt)) {
       throw new Error('There is no interrupted or prior run to continue yet.');
     }
 
-    return await submit({
+    return await this.submit({
       ...input,
       prompt: input.prompt ?? session.lastContinuePrompt ?? '',
     });
   }
 
-  return {
-    submit,
-    continue: continueTurn,
-    clearLease(input) {
-      clearConversationTurnLease(config.sessionStoragePath, input.sessionId, input.owner);
-    },
-  };
+  clearLease(input: ClearConversationTurnLeaseInput): void {
+    clearConversationTurnLease(this.config.sessionStoragePath, input.sessionId, input.owner);
+  }
 }
