+++ src/server/index.ts
@@ -17,7 +17,7 @@ export type { HeddleServerListenOptions, HeddleServerOptions } from './types.js'
 export { appRouter, type AppRouter } from './router.js';
 export { createHeddleServerApp } from './app.js';
 export { createServerLogger } from './logger.js';
-export { projectChatSessionView } from './features/control-plane/services/chat-sessions.js';
+export { ControlPlaneChatSessionPresenter } from './features/control-plane/controllers/chat-session-presenter.js';
 
 export async function listenHeddleDaemon(options: HeddleServerListenOptions): Promise<void> {
   const serveAssets = options.serveAssets !== false;
